+++ src/Pages/Cart/Cart.jsx
@@ -1,14 +1,13 @@
 import { useState, useEffect } from 'react';
-import { Checkout, Loaders } from '../../Components';
-import { HorizontalCard } from '../../Components/Horizontal_Card/HorizontalCard';
-import { useAuth, useCart } from '../../Context';
-import PriceContainer from './Components/Price Container/PriceContainer';
 import { Link } from 'react-router-dom';
+import axios from 'axios';
+import { toast } from 'react-toastify';
+import { Checkout, Loaders, HorizontalCard } from '../../Components';
+import { PriceContainer } from './Components/Price Container/PriceContainer';
+import { useAuth, useCart } from '../../Context';
 import { fetchCartProducts } from '../../Utils';
 import { emptyCart } from '../../Assets/images';
 import styles from './Cart.module.css';
-import axios from 'axios';
-import { toast } from 'react-toastify';
 
 const Cart = () => {
 	const { authState } = useAuth();
