+++ src/Pages/Home/Components/Categories/Categories.jsx
@@ -38,4 +38,4 @@ const Categories = () => {
 	);
 };
 
-export default Categories;
+export { Categories };
