+++ src/Pages/Home/Components/Footer/Footer.jsx
@@ -45,4 +45,4 @@ const Footer = () => {
 	);
 };
 
-export default Footer;
+export { Footer };
