+++ src/Pages/Home/Components/index.js
@@ -1,0 +1,3 @@
+export { Carousel } from './Carousel/Carousel';
+export { Categories } from './Categories/Categories';
+export { Footer } from './Footer/Footer';
