+++ src/Pages/Home/Home.jsx
@@ -1,9 +1,7 @@
 import { Link } from 'react-router-dom';
-import Carousel from './Components/Carousel/Carousel';
-import Categories from './Components/Categories/Categories';
-import Footer from './Components/Footer/Footer';
-import styles from './Home.module.css';
+import { Carousel, Categories, Footer } from './Components';
 import { horizon } from '../../Assets/images';
+import styles from './Home.module.css';
 
 const Home = () => {
 	return (
